requirements.txt
# APE Core 필수 의존성
fastapi>=0.95.0,<1.0.0
uvicorn>=0.21.1,<0.22.0
pydantic>=1.10.7,<2.0.0
requests>=2.28.2,<3.0.0
chromadb>=0.4.13,<0.5.0
python-dotenv>=1.0.0,<2.0.0
sentence-transformers>=2.2.2,<3.0.0
python-multipart>=0.0.6,<0.1.0
tqdm>=4.65.0,<5.0.0

# 벡터 데이터베이스 및 임베딩
numpy>=1.22.0,<2.0.0
torch>=2.0.0,<3.0.0

# AI/ML 관련 라이브러리
langgraph>=0.0.10,<0.1.0
langchain>=0.0.200,<0.1.0
typing-extensions>=4.5.0,<5.0.0

# 개발용 의존성
pytest>=7.0.0,<8.0.0
black>=23.0.0,<24.0.0
isort>=5.12.0,<6.0.0
mypy>=1.0.0,<2.0.0